First cut at basic dependency resolution
Based on an algorithm described in this blog post:

  http://www.electricmonk.nl/log/2008/08/07/dependency-resolving-algorithm/

Chosen for its simplicity, given that extreme efficiency is unlikely to
be needed with the anticipated working set sizes (and if it ever does
become an issue, a more efficient algorithm can be investigated).

Note: the blog post in question makes the issue seem more simple than it
really is as it says, "we'll start with node 'A'", which just happens to
be the only node in the graph which will result in the dependencies
being fully resolved. Starting from any other node will result in an
incomplete dependency graph. So, we need to iterate over all nodes,
applying the algorithm to each in turn, to ensure that we get a complete
graph.

Other algorithms are mentioned here:

  http://en.wikipedia.org/wiki/Topological_sorting

Signed-off-by: Wincent Colaiuta <win@wincent.com>